add editable window to set rolling average duration
Added an editable window to the plugin that allows the user to set rolling average duration. Required making a new class member to update the rolling window size to separate its reset from updating the bandwidth.  It takes one width of the window's duration for the signal to return to a stable value because it reinitializes the window

Also removed the last remnants of the crossing detector plugin